Allow grouped/namespaced exploration of cli commands and scope by service token
