+++ lib/toknsmith/tokens.rb
@@ -1,55 +1,16 @@
 # frozen_string_literal: true
 
 require "thor"
-require_relative "client"
-require "dotenv/load"
+require_relative "service_tokens"
+require_relative "oauth_tokens"
 
 module Toknsmith
-  # Handles CLI token storage for external services (e.g., GitHub, GitLab, etc.)
+  # Parent CLI command to route token subcommands
   class Tokens < Thor
-    desc "store SERVICE", "Store a token for a given service"
-    option :token, required: true, aliases: "-t", desc: "The actual token value"
-    option :note, aliases: "-n", desc: "Optional note for the token"
-    option :expires_in, aliases: "-e", desc: "Optional expiration (e.g. '30d', '12h')"
+    desc "service SUBCOMMAND", "Manage service tokens (e.g. GitHub PATs)"
+    subcommand "service", ServiceTokens
 
-    def store(service)
-      client = Client.new
-      response = client.store_service_token(service, options[:token], options[:note], options[:expires_in])
-
-      if response
-        print_token_result(service, response)
-      else
-        puts "❌ Failed to store token."
-      end
-    rescue StandardError => e
-      puts "Error: #{e.message}"
-    end
-
-    desc "list", "List stored tokens for your team"
-    def list
-      client = Client.new
-      tokens = client.list_service_tokens
-
-      if tokens.any?
-        puts "🔐 Stored Tokens:"
-        puts "ID\tService\t\tExpires\t\t\tNote"
-        tokens.each do |t|
-          puts "#{t["id"]}\t#{t["service_name"]}\t#{t["expires_at"] || "Never"}\t#{t["note"] || "-"}"
-        end
-      else
-        puts "😴 No tokens stored yet."
-      end
-    rescue StandardError => e
-      puts "Error: #{e.message}"
-    end
-
-    private
-
-    def print_token_result(service, response)
-      puts "🔐 Token for #{service} stored securely."
-      puts "ID: #{response["id"]}"
-      puts "Note: #{response["note"]}" if response["note"]
-      puts "Expires: #{response["expires_at"]}" if response["expires_at"]
-    end
+    desc "oauth SUBCOMMAND", "Manage OAuth tokens (via provider flows)"
+    subcommand "oauth", OauthTokens
   end
 end
